// src/app/model/contrato.ts
export interface Contrato {
    id?: number;
    nombre: string;
    direccion: string;
    telefono: string;
    categoria: number;
    proveedor: number;
    usuario: number;
    fecha: string;
    hora?: string;
    minutos?: string;
    convencional?: string;
    descripcion?: string;
    forma?: number;
    estado: string;
    referencia?: string;
    created_at?: string;
    updated_at?: string;
}